% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{mse1.extract_prm_from_mseres}
\alias{mse1.extract_prm_from_mseres}
\title{mse1.extract_prm_from_mseres}
\usage{
mse1.extract_prm_from_mseres(mseres, stepidx)
}
\description{
mse1.extract_prm_from_mseres
}
