% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{mse1.plot_graph}
\alias{mse1.plot_graph}
\title{mse1.plot_graph}
\usage{
mse1.plot_graph(
  mseprm,
  vertices_type = c("Patch", "Cell", "Chromosome", "Plasmid", "Gene"),
  add_archetype = TRUE,
  edges_type = c("inclusion", "plasmid_range", "diffusion"),
  color_map = c(Patch = "#FF7043", Cell = "#FFEE58", Chromosome = "#66BB6A", Plasmid =
    "#42A5F5", Gene = "#EF5350"),
  layout = c("hr", "fr", "free")[1]
)
}
\description{
mse1.plot_graph
}
